sftp: Prompt for password, don't terminate on SIGINT
This is a follow-up on 6f59e8c504acf593930ac7714963b269478db830, which
runs the `ssh` in its own process group and selects that process group
as the foreground group. After the sftp connection is established,
restic switches back to the previous foreground process group.

This allows `ssh` to prompt for the password, but it won't receive
the interrupt signal (SIGINT, ^C) later on, because it is not in the
foreground process group any more, allowing a clean tear down.