lock: Ignore invalid lock file
This commit fixes a bug introduced in
eca3751bc0e4d29505a3db7a24acce9341b1897d: When an invalid lock is
encountered (e.g. if the file is empty), the code used to ignore that,
but now returns the error.

Now, invalid files are ignored for the normal lock check, and removed
when `restic unlock --remove-all` is run.

Closes #1652